+++ CHANGELOG.md
@@ -7,7 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+### Added
+- Added new `--min-monthly-fee` and `--max-monthly-fee` options to filter flats by
+  monthly fee. The default is no minimum or maximum monthly fee.
+- Added new `--cache/--no-cache` flag, which allows you to disable the cache. The
+  default behaviour is still to use the cache, but you can disable it by using the
+  `--no-cache` flag. This is useful if you want to see all the results, and not just the
+  new ones. The cache is stored in the `.boligping_cache` file in the current directory.
 
+### Changed
+- Changed the `--query` (`-q`) option to now search for the selected keywords in the
+  description of the flat, rather than use Boligsiden.dk's own keywords, since many
+  flats do not use these keywords. Remember that you can specify multiple queries with,
+  e.g., `bolig-ping <other-arguments> -q badekar -q altan`.
 
 
 ## [v1.2.0] - 2025-04-04